+++ CHANGELOG.md
@@ -4,11 +4,20 @@ Please use [this](https://docs.gitlab.com/ee/development/changelog.html) documen
 
 ## [unreleased]
 
+### Added
+
+* config: Added `custom` section to store arbitrary user data.
+
 ### Fixed
 
 * config: Fixed default SQLite path (`:memory:`).
 * tzkt: Fixed pagination in several getter methods.
 * tzkt: Fixed data loss when `skip_history` option is enabled.
+
+### Removed
+
+* config: Removed dummy `advanced.oneshot` flag.
+* cli: Removed `docker init` command.
 
 ## 5.0.0-rc2 - 2022-03-13